# Improve performance of get_block_theme_folders