--- a/src/wp-includes/block-template-utils.php
+++ b/src/wp-includes/block-template-utils.php
@@ -37,7 +37,8 @@ if ( ! defined( 'WP_TEMPLATE_PART_AREA_UNCATEGORIZED' ) ) {
  * }
  */
 function get_block_theme_folders( $theme_stylesheet = null ) {
-	$theme = wp_get_theme( $theme_stylesheet );
+	$theme_name = null === $theme_stylesheet ? get_stylesheet() : $theme_stylesheet;
+	$theme      = wp_get_theme( $theme_name );
 	return $theme->get_block_template_folders();
 }