--- a/src/wp-includes/block-template-utils.php
+++ b/src/wp-includes/block-template-utils.php
@@ -36,8 +36,15 @@ if ( ! defined( 'WP_TEMPLATE_PART_AREA_UNCATEGORIZED' ) ) {
  *     @type string $wp_template_part Theme-relative directory name for block template parts.
  * }
  */
-function get_block_theme_folders( $theme_stylesheet = '' ) {
+function get_block_theme_folders( $theme_stylesheet = null ) {
 	$theme = wp_get_theme( $theme_stylesheet );
+	if ( ! $theme->exists() ) {
+		// Return the default folders if the theme doesn't exist.
+		return array(
+			'wp_template'      => 'templates',
+			'wp_template_part' => 'parts',
+		);
+	}
 	return $theme->get_block_template_folders();
 }