--- a/tests/phpunit/data/themedir1/block-theme-child-deprecated-path/style.css
+++ b/tests/phpunit/data/themedir1/block-theme-child-deprecated-path/style.css
@@ -1,0 +1,8 @@
+/*
+Theme Name: Block Theme Child with deprecated Path
+Theme URI: https://wordpress.org/
+Description: For testing purposes only.
+Template: block-theme
+Version: 1.0.0
+Text Domain: block-theme-child-deprecated-path
+*/