… tests/phpunit/data/themedir1/block-theme-child-deprecated-path/block-template-parts/small-header.html
… tests/phpunit/data/themedir1/block-theme-child-deprecated-path/templates/index.html
… tests/phpunit/data/themedir1/block-theme-child-deprecated-path/theme.json
… tests/phpunit/tests/block-template.php
… tests/phpunit/tests/theme/themeDir.php
… tests/phpunit/tests/theme/wpTheme.php