Trac ticket: [https://core.trac.wordpress.org/ticket/58319](https://core.trac.wordpress.org/ticket/58319)

----
Benchmark for TT3 site showed a considerable performance improvement after cache implementation:

| Metric                       | Before Cache | After Cache |
|------------------------------|--------------|-------------|
| Response Time (median)       | 160.65       | 151.17      |
| wp-before-template (median)  | 61.11        | 58.18       |
| wp-template (median)         | 95.42        | 90.1        |
| wp-total (median)            | 157.77       | 148.39      |


closes #4456

---
**This Pull Request is for code review only. Please keep all other discussion in the Trac ticket. Do not merge this Pull Request. See [GitHub Pull Requests for Code Review](https://make.wordpress.org/core/handbook/contribute/git/github-pull-requests-for-code-review/) in the Core Handbook for more details.**