@spacedmonkey 2023-09-14
@kt-12 Please check https://github.com/WordPress/wordpress-develop/pull/3629/files to check how caching was added here.